- Moved JSON-based auto to java api, renamed as EasyAuto - Modified sleep function to use seconds, removed pointless error throwing
