--- TeamCode/src/main/resources/JSONAutos/exampleAuto.json
@@ -1,16 +1,0 @@
-[
-  {
-    "Action Type" : "Lateral Movement",
-    "Distance X" : 20.0,
-    "Distance Y" : 10.0,
-    "Time" : 1.5
-  }, {
-    "Action Type" : "Sleep",
-    "Time" : 2.5
-  }, {
-    "Action Type" : "Lateral Movement",
-    "Distance X" : -20.0,
-    "Distance Y" : -10.0,
-    "Time" : 1.5
-  }
-]
